+++ fittrack_mobile_app/lib/widgets/meal_view_widget.dart
@@ -5,7 +5,7 @@ import '../styles/colors.dart';
 
 class MealViewWidget extends StatelessWidget {
   final String foodName;
-  final int kcal;
+  final double kcal;
   final VoidCallback onDelete;
 
   const MealViewWidget({
